Add files via upload
Build a chatbot by means of an LSTM encoder and a decoder LSTM. Prepare input and output datasets by tokenizing and padding human-robot conversations from Kaggle then creating an embedding matrix with reference to GloVe. The human part of the input set is fed to the encoder LSTM after an Embedding layer to reduce dimension and the output hidden states and cells go to the decoder LSTM along with the input target sentences. The decoder aims to learn how to generate an outcome like the output one. The prediction phase requires only training the decoder from the hidden and cell state from the input sequence and the start of the word for the target sequence to bring out the corresponding target word, the cell state and the hidden state,which, together with the predicted word, will be used as input to the next LSTM unit. Finally, to generate a response to the input, we need to translate the output matrix into words and voila, un Chatbot.